refactor: folder renaming
