+++ src/draft-assistant/DraftAssistant.ts
@@ -1,4 +1,4 @@
-import type { LabelId, LabeledRule } from "@/assertion";
+import type { LabelId, LabeledRule } from "@/rule";
 import type { ModelFromContainer, DraftViewer } from "@/types";
 
 /**
