+++ src/draft-assistant/SectionDraftAssistant.ts
@@ -1,7 +1,7 @@
 import { DraftAssistant } from "./DraftAssistant";
-import { RulesBroken } from "@/assertion";
+import { RulesBroken } from "@/rule";
 
-import type { LabelId, LabeledRule } from "@/assertion";
+import type { LabelId, LabeledRule } from "@/rule";
 import type { ModelFromContainer, AssistantsIn } from "@/types";
 
 export type CreationClosure<Model, ComposedModels extends unknown[]> = (...models: ComposedModels) => Model;
