+++ src/draft-publisher/DraftPublisher.ts
@@ -2,7 +2,7 @@ import { EventEmitter } from "events";
 
 import type { DraftAssistant } from "@/draft-assistant";
 import type { DraftViewer } from "@/types";
-import type { LabeledRule } from "@/assertion";
+import type { LabeledRule } from "@/rule";
 
 type PublisherEvents<Model = unknown> =
   | { "draft:updated": [Model] }
