+++ src/rule-requirements/lists/ListsConditions.test.ts
