+++ src/rule-requirements/lists/ListsConditions.ts
