+++ src/rule/Inquiry.test.ts
