+++ src/rule/index.ts
@@ -1,4 +1,6 @@
 export * from "./Assertion";
+export * from "./Inquiry";
+export * from "./Rule";
 export * from "./RuleLabel";
 export * from "./RulesBroken";
 export * from "./Ruleset";
